requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
cssbeautifier==1.15.4
djlint==1.36.4
dnspython==2.8.0
EditorConfig==0.17.1
Faker==38.2.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.12.1
MarkupSafe==3.0.3
numpy==2.3.5
pandas==2.3.3
pathspec==0.12.1
pymongo==4.15.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
six==1.17.0
SQLAlchemy==2.0.44
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
